feat: separate mailerService into dedicated email handlers and add authRoute
- Split `mailerService.js` into:
  - `sendChangePassword.js`
  - `sendDeleteAccountConfirm.js`
  - `sendEmailRegister.js`
  - `sendForgotPassword.js`
- Added `authRoute.js` for authentication-related endpoints
- Improved project structure for better maintainability and clarity
