+++ backend/src/libs/sendChangePassword.js
@@ -1,0 +1,50 @@
+import { client } from '../config/mailgunConfig.js'
+import dotenv from 'dotenv'
+
+dotenv.config()
+
+export const sendChangePassword = async (user) => {
+  try {
+    const html = `
+      <div style="font-family: Arial, sans-serif; max-width: 600px; margin: auto; background: #fff; border: 1px solid #e0e0e0; border-radius: 8px; padding: 24px; color: #333;">
+        <h2 style="color: #A48DC7; text-align: center;">Password Changed Successfully</h2>
+
+        <p>Hello <strong>${user.name}</strong>,</p>
+
+        <p>Your password for <strong>TerraQuake API</strong> has been successfully changed.</p>
+
+        <p>If you did not make this change, please reset your password immediately or contact our support team.</p>
+
+        <div style="text-align: center; margin: 24px 0;">
+          <a href="${process.env.FRONTEND_PRODUCTION || process.env.FRONTEND_DEVELOPMENT}" 
+             style="background: #A48DC7; color: #fff; padding: 12px 20px; text-decoration: none; border-radius: 6px; font-weight: bold;">
+            Go to TerraQuake API
+          </a>
+        </div>
+
+        <hr style="margin: 32px 0; border: none; border-top: 1px solid #ddd;" />
+
+        <p style="font-size: 0.9em; color: #666;">
+          Need help or have questions? Contact our support team at 
+          <a href="mailto:terraquakeapi@gmail.com">terraquakeapi@gmail.com</a>.
+        </p>
+
+        <p style="font-weight: bold; font-size: 1.1em; color: #333; text-align: center;">
+          The <span style="color: #A48DC7;">TerraQuake API</span> Team
+        </p>
+      </div>
+    `
+    const result = await client.messages.create(process.env.MAILGUN_DOMAIN, {
+      from: `TerraQuake API <postmaster@${process.env.MAILGUN_DOMAIN}>`,
+      to: user.email,
+      subject: '🔒 Your TerraQuake API Password Has Been Changed',
+      html
+    })
+
+    console.log('Change password email sent:', result)
+    return result
+  } catch (error) {
+    console.error('Error sending change password email:', error)
+    throw error
+  }
+}
